cleanup, comments
