+++ python_modules/libraries/dagster-airbyte/dagster_airbyte/__init__.py
@@ -3,7 +3,14 @@ from dagster._core.utils import check_dagster_package_version
 try:
     import dagster_managed_stacks
 
-    from . import managed
+    from .managed import (
+        AirbyteConnection,
+        AirbyteDestination,
+        AirbyteManagedStackReconciler,
+        AirbyteSource,
+        AirbyteSyncMode,
+    )
+
 except ImportError:
     pass
 
@@ -14,13 +21,7 @@ from .asset_defs import (
 )
 from .ops import airbyte_sync_op
 from .resources import AirbyteResource, AirbyteState, airbyte_resource
-from .types import (
-    AirbyteConnection,
-    AirbyteDestination,
-    AirbyteOutput,
-    AirbyteSource,
-    AirbyteSyncMode,
-)
+from .types import AirbyteOutput
 from .version import __version__
 
 check_dagster_package_version("dagster-airbyte", __version__)
