+++ python_modules/libraries/dagster-airbyte/dagster_airbyte/managed/__init__.py
@@ -1,0 +1,2 @@
+from .reconciliation import AirbyteManagedStackReconciler
+from .types import AirbyteConnection, AirbyteDestination, AirbyteSource, AirbyteSyncMode
