+++ python_modules/libraries/dagster-airbyte/dagster_airbyte_tests/example_stacks/empty_airbyte_stack.py
@@ -1,0 +1,19 @@
+import os
+
+from dagster_airbyte import (
+    AirbyteManagedStackReconciler,
+    airbyte_resource,
+)
+
+airbyte_instance = airbyte_resource.configured(
+    {
+        "host": os.getenv("AIRBYTE_HOSTNAME", "localhost"),
+        "port": os.getenv("AIRBYTE_PORT", "80"),
+    }
+)
+
+
+reconciler = AirbyteManagedStackReconciler(
+    airbyte=airbyte_instance,
+    connections=[],
+)
