+++ python_modules/libraries/dagster-airbyte/dagster_airbyte_tests/example_stacks/example_airbyte_stack.py
@@ -1,20 +1,13 @@
 import os
-from abc import abstractmethod
-from typing import Optional, Set
 
 from dagster_airbyte import (
     AirbyteConnection,
     AirbyteDestination,
+    AirbyteManagedStackReconciler,
     AirbyteSource,
     AirbyteSyncMode,
     airbyte_resource,
-    load_assets_from_airbyte_instance,
-    load_assets_from_airbyte_project,
 )
-from dagster_airbyte.managed import AirbyteManagedStackReconciler
-
-from dagster import build_init_resource_context, repository, with_resources
-from dagster._core.definitions.resource_requirement import ResourceAddable
 
 airbyte_instance = airbyte_resource.configured(
     {
@@ -22,14 +15,6 @@ airbyte_instance = airbyte_resource.configured(
         "port": os.getenv("AIRBYTE_PORT", "80"),
     }
 )
-
-# pokeapi_source = AirbyteSource(
-#     name="pokeapi-source",
-#     source_type="PokeAPI",
-#     source_configuration={
-#         "pokemon_name": "snorlax",
-#     },
-# )
 
 
 local_json_source = AirbyteSource(
@@ -46,11 +31,11 @@ local_json_source = AirbyteSource(
 local_json_destination = AirbyteDestination(
     name="local-json-output",
     destination_type="Local JSON",
-    destination_configuration={"destination_path": "./d"},
+    destination_configuration={"destination_path": "/local/destination_file.json"},
 )
 
 
-pokeapi_to_local_json_conn = AirbyteConnection(
+local_json_conn = AirbyteConnection(
     name="local-json-conn",
     source=local_json_source,
     destination=local_json_destination,
@@ -64,6 +49,6 @@ pokeapi_to_local_json_conn = AirbyteConnection(
 reconciler = AirbyteManagedStackReconciler(
     airbyte=airbyte_instance,
     connections=[
-        pokeapi_to_local_json_conn,
+        local_json_conn,
     ],
 )
